Added activity sync.

+++ Scozzard.Model/Activity.cs
@@ -12,7 +12,7 @@ namespace Scozzard.Model
         public DateTime StartTime { get; set; }
         public string Description { get; set; }
         public string ImageUrl { get; set; }
-        public int XboxUserID { get; set; }
+        public long XboxUserID { get; set; }
         public XboxUser XboxUser { get; set; }
     }
 }
